bip-taro-addr: convert internal addr bytes to TLV format
In this commit, we convert the address payload itself to be a series of
TLV fields. This makes it easier to update in the future, and also lets
the receiver express that certain requirements MUST be understood if a
transfer is to be completed.

We also carry over the recent changes w.r.t the `asset_key_family` field
an its role in ensure that assets with ongoing issuance can be nested in
the MS-SMT tree. We also lift the asset type into the address itself, as
combined with the aforementioned change, this enables sending
collectibles on chain more easily.