bip-taro-proof-file: unroll proof file to nested TLV, add exclusion proofs
In this commit, we catch uyp the proof file format to be in line with
the core Taro BIP w.r.t the series of exclusion proofs needed to prove
commitment uniqueness, as well as the extra tapscript pre-image
information need to prove that a commitment is in the proper place in
that taproot tree.

Along the way we split everything up into a series of nested TLVs, and
also add an additional field for a set of proofs for each input split
that may be created during transfers. The TODO left in place still
stands, as if we add that input MMR, then verifies can use ot to select
random state transitions to verify. A related note is that the set of
input split proofs can also optionally be verified as a false Proof can
end up burning coins, so verifies can make a call w.r.t the expected
value destroyed/created with a fake proof attempt.